+++ Plugin.cs
@@ -8,16 +8,16 @@ using Assets.Scripts.Inventory__Items__Pickups.Items;
 
 namespace MegaBonkMod;
 
-[BepInPlugin("com.megabonk.mod", "MegaBonk Mod", "1.3.14")]
+[BepInPlugin("com.megabonk.mod", "MegaBonk Mod", "1.3.15")]
 public class Plugin : BasePlugin
 {
     internal static string LeaderboardServer = "http://megabonkcommunitypatch.duckdns.org:9000";
-    internal const  string ModVersion        = "1.3.14";
+    internal const  string ModVersion        = "1.3.15";
 
     internal const bool PatchGrandmasTonic = true;
 
     internal static readonly HashSet<EItem> ActiveUncappedItems   = new() { EItem.Anvil, EItem.OverpoweredLamp, EItem.ZaWarudo };
-    internal static readonly HashSet<EItem> ForcedPoolItems       = new() { EItem.Battery, EItem.Skuleg, EItem.OldMask, EItem.BrassKnuckles, EItem.DemonicBlood, EItem.IdleJuice, EItem.SuckyMagnet };
+    internal static readonly HashSet<EItem> ForcedPoolItems       = new() { EItem.Battery, EItem.Skuleg, EItem.OldMask, EItem.BrassKnuckles, EItem.DemonicBlood, EItem.IdleJuice, EItem.SuckyMagnet, EItem.Key, EItem.EchoShard };
     internal static readonly HashSet<int>   BlacklistedStats       = new() { 1,2,3,4,5,10,11,24 };
     internal static readonly HashSet<int>   BlacklistedShrineStats = new() { 1,2,3,4,5,10,11,24 };
     internal static readonly List<int>      FullStatPool           = new();